🐛  remove the obsolete recommendation component
